// src/thirdparty/Genshin-Impact-Miliastra-Wonderland-Code-Node-Editor-Pack/protobuf/gia.proto.ts
export interface Root {
  /** Index = 1 */
  graph: GraphUnit
  /** Index = 2 */
  accessories: GraphUnit[]
  /** Index = 3 */
  filePath: string
  /** Index = 5 */
  gameVersion: string
}
export interface GraphUnit {
  /** Index = 1 */
  id: GraphUnit_Id
  /** Index = 2 */
  relatedIds: GraphUnit_Id[]
  /** Index = 3 */
  name: string
  /** Index = 5 */
  which: GraphUnit_Which
  /** Index = 13
   *
   * One of the field: graphType
   */
  graph?: NodeGraphWrapper
  /** Index = 14
   *
   * One of the field: graphType
   */
  compositeDef?: CompositeDefWrapper
  /** Index = 22
   *
   * One of the field: graphType
   */
  structureDef?: StructureDefWrapper
}
export interface GraphUnit_Id {
  /** Index = 2 */
  class: GraphUnit_Id_Class
  /** Index = 3 */
  type: GraphUnit_Id_Type
  /** Index = 4 */
  id: number
}
export const GraphUnit_Id_Class = {
  Unknown1: 0,
  Node: 1,
  Basic: 5,
  AffiliatedNode: 23
} as const
export type GraphUnit_Id_Class = (typeof GraphUnit_Id_Class)[keyof typeof GraphUnit_Id_Class]
export const GraphUnit_Id_Type = {
  ServerGraph: 0,
  ClientGraph: 3,
  StructureDefinition: 15
} as const
export type GraphUnit_Id_Type = (typeof GraphUnit_Id_Type)[keyof typeof GraphUnit_Id_Type]
export const GraphUnit_Which = {
  Unknown: 0,
  EntityNode: 9,
  BooleanFilter: 10,
  Skills: 11,
  CompositeGraph: 12,
  StatusNode: 22,
  ClassNode: 23,
  StructureDefinition: 29,
  ItemNode: 46,
  IntegerFilter: 47
} as const
export type GraphUnit_Which = (typeof GraphUnit_Which)[keyof typeof GraphUnit_Which]
export interface NodeGraphWrapper {
  /** Index = 1 */
  inner: NodeGraphWrapper_InnerWrapper
}
export interface NodeGraphWrapper_InnerWrapper {
  /** Index = 1 */
  graph: NodeGraph
}
export interface CompositeDefWrapper {
  /** Index = 1 */
  inner: CompositeDefWrapper_InnerWrapper
}
// ...
export interface NodeGraph_Id {
  /** Index = 1 */
  class: NodeGraph_Id_Class
  /** Index = 2 */
  type: NodeGraph_Id_Type
  /** Index = 3 */
  kind: NodeGraph_Id_Kind
  /** Index = 5 */
  id: number
}
export const NodeGraph_Id_Class = {
  Unknown1: 0,
  UserDefined: 10000,
  SystemDefined: 10001
} as const
export type NodeGraph_Id_Class = (typeof NodeGraph_Id_Class)[keyof typeof NodeGraph_Id_Class]
export const NodeGraph_Id_Type = {
  Unknown2: 0,
  BasicNode: 20000,
  BooleanFilter: 20001,
  Skills: 20002,
  StatusNode: 20003,
  ClassNode: 20004,
  ItemNode: 20005,
  IntegerFilter: 20006
} as const
export type NodeGraph_Id_Type = (typeof NodeGraph_Id_Type)[keyof typeof NodeGraph_Id_Type]
export const NodeGraph_Id_Kind = {
  Unknown3: 0,
  NodeGraph: 21001,
  CompositeGraph: 21002,
  SysCall: 22000,
  SysGraph: 22001
} as const
export type NodeGraph_Id_Kind = (typeof NodeGraph_Id_Kind)[keyof typeof NodeGraph_Id_Kind]
export interface Comments {
  /** Index = 1 */
  content: string
  /** Index = 2 */
  x?: number
  /** Index = 3 */
  y?: number
}
export interface GraphAffiliation {
  /** Index = 1 */
  info: GraphAffiliation_Info
  /** Index = 2 */
  type: GraphAffiliation_Type
}
export interface GraphAffiliation_Info {
  /** Index = 1 */
  source: NodeGraph_Id
  /** Index = 2 */
  typeXxx: number
  /** Index = 3 */
  alwaysOneXxxx: number
  /** Index = 100
   *
   * One of the field: extend
   */
  structId?: VarBase_ItemType_StructItem
}
export interface GraphAffiliation_Type {
  /** Index = 1 */
  type: number
}
export interface StructureDefWrapper {
  /** Index = 1 */
  def: StructureDefWrapper_StructureDef
}
// ...
export interface VarBase {
  /** Index = 1 */
  class: VarBase_Class
  /** Index = 2 */
  alreadySetVal: boolean
  /** Index = 4 */
  itemType?: VarBase_ItemType
  /** Index = 5 */
  structInfo?: VarBase_StructInfo
  /** Index = 101
   *
   * One of the field: baseValues
   */
  bId?: IdBaseValue
  /** Index = 102
   *
   * One of the field: baseValues
   */
  bInt?: IntBaseValue
  /** Index = 104
   *
   * One of the field: baseValues
   */
  bFloat?: FloatBaseValue
  /** Index = 105
   *
   * One of the field: baseValues
   */
  bString?: StringBaseValue
  /** Index = 106
   *
   * One of the field: baseValues
   */
  bEnum?: EnumBaseValue
  /** Index = 107
   *
   * One of the field: baseValues
   */
  bVector?: VectorBaseValue
  /** Index = 108
   *
   * One of the field: baseValues
   */
  bStruct?: StructBaseValue
  /** Index = 109
   *
   * One of the field: baseValues
   */
  bArray?: ArrayBaseValue
  /** Index = 110
   *
   * One of the field: baseValues
   */
  bConcreteValue?: ConcreteBaseValue
  /** Index = 111
   *
   * One of the field: baseValues
   */
  bMapPair?: MapPairBaseValue
  /** Index = 112
   *
   * One of the field: baseValues
   */
  bMap?: MapBaseValue
}
// ...
export interface VarBase_ItemType_StructItem {
  /** Index = 1 */
  structId: number
}
// ...
export interface GraphNode {
  /** Index = 1 */
  nodeIndex: number
  /** Index = 2 */
  genericId: NodeProperty
  /** Index = 3 */
  concreteId?: NodeProperty
  /** Index = 4 */
  pins: NodePin[]
  /** Index = 5 */
  x: number
  /** Index = 6 */
  y: number
  /** Index = 7 */
  comments?: Comments
  /** Index = 10 */
  usingStruct: GraphAffiliation_Info[]
}
export interface NodeProperty {
  /** Index = 1 */
  class: NodeGraph_Id_Class
  /** Index = 2 */
  type: NodeProperty_Type
  /** Index = 3 */
  kind: NodeGraph_Id_Kind
  /** Index = 5 */
  nodeId: number
}
export const NodeProperty_Type = {
  Unknown: 0,
  Server: 20000,
  Filter: 20001,
  Skill: 20002
} as const
export type NodeProperty_Type = (typeof NodeProperty_Type)[keyof typeof NodeProperty_Type]
export interface NodePin {
  /** Index = 1 */
  i1: NodePin_Index
  /** Index = 2 */
  i2: NodePin_Index
  /** Index = 3 */
  value: VarBase
  /** Index = 4 */
  type: number
  /** Index = 5 */
  connects: NodeConnection[]
  /** Index = 6 */
  clientExecNode?: NodePin_Index
  /** Index = 7 */
  compositePinIndex?: number
}
export interface NodePin_Index {
  /** Index = 1 */
  kind: NodePin_Index_Kind
  /** Index = 2 */
  index: number
  /** Index = 100
   *
   * One of the field: clientExecNode
   */
  nodeId?: NodePin_Index_Id
}
export const NodePin_Index_Kind = {
  Unknown: 0,
  InFlow: 1,
  OutFlow: 2,
  InParam: 3,
  OutParam: 4,
  ClientExecNode: 5
} as const
export type NodePin_Index_Kind = (typeof NodePin_Index_Kind)[keyof typeof NodePin_Index_Kind]
export interface NodePin_Index_Id {
  /** Index = 1 */
  id: number
}
export interface NodeConnection {
  /** Index = 1 */
  id: number
  /** Index = 2 */
  connect: NodePin_Index
  /** Index = 3 */
  connect2: NodePin_Index
}